requirements.txt
Pillow~=8.1.0
antlr4-python3-runtime~=4.9.1
python-telegram-bot~=13.1